Refactor sibling migrators into partials of new top-level migrator
